+++ lib/braintree/payment_method_gateway.rb
@@ -172,6 +172,18 @@ module Braintree
         {:options => options},
         {:billing_address => billing_address_params}
       ]
+      signature << {
+        :three_d_secure_pass_thru => [
+          :eci_flag,
+          :cavv,
+          :xid,
+          :three_d_secure_version,
+          :authentication_response,
+          :directory_response,
+          :cavv_algorithm,
+          :ds_transaction_id,
+        ]
+      }
 
       case type
       when :create