+++ lib/braintree/payment_method_nonce.rb
@@ -19,12 +19,14 @@ module Braintree
     attr_reader :nonce
     attr_reader :three_d_secure_info
     attr_reader :type
+    attr_reader :authentication_insight
 
     def initialize(gateway, attributes) # :nodoc:
       @gateway = gateway
       @nonce = attributes.fetch(:nonce)
       @type = attributes.fetch(:type)
       @details = attributes.fetch(:details)
+      @authentication_insight = attributes.fetch(:authentication_insight, nil)
       @three_d_secure_info = ThreeDSecureInfo.new(attributes[:three_d_secure_info]) if attributes[:three_d_secure_info]
       @bin_data = BinData.new(attributes[:bin_data]) if attributes[:bin_data]
     end